1. 更改ECPM类型

+++ adsdk/src/main/java/com/inner/adsdk/adloader/listener/ISdkLoader.java
@@ -83,5 +83,5 @@ public interface ISdkLoader {
 
     public int getBannerSize();
 
-    public int getEcpm();
+    public double getEcpm();
 }
